% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy2clipboard.R
\name{copy_plot_to_clipboard}
\alias{copy_plot_to_clipboard}
\title{Copy Plots to Clipboard Function}
\usage{
copy_plot_to_clipboard(filepath = NULL)
}
\arguments{
\item{filepath}{filepath.}
}
\value{
Nothing. Only copies image to clipboard.
}
\description{
This function takes an image file and copies the image into the clipboard. It uses nircmd so it is only available on windows.
}
\examples{
filepath <- "Plots/foo.png"
copy_plot_to_clipboard(filepath)
}
\keyword{Clipboard,}
\keyword{Plot}
