% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_my_theme.R
\name{set_my_theme}
\alias{set_my_theme}
\title{Set theme for my plots}
\usage{
set_my_theme()
}
\value{
Nothing.
}
\description{
This function sets the theme I use for most of my plots
}
\examples{
set_my_theme()
}
\keyword{Plot}
\keyword{Theme,}
